ENH: Increase test accuracy

+++ mutual/src/nl/hanze/research/mutual/PrintResults.java
@@ -2,7 +2,7 @@ package nl.hanze.research.mutual;
 
 public class PrintResults {
 
-    public static void of(String caption, Long[] measurements) {
+    public static void of(String caption, Double[] measurements) {
         String message = caption +
                 "\n-\taverage: " +
                 Average.of(measurements) +
